Reorganized import to keep example working as a self-contained unit.
NOTE: Remember to place this in a directory/folder within OpenRecommender so that the paths are correct. Suggest using something like "OPEN_RECOMMENDER_PATH/client/php" and likewise for other languages. There are still no hard library dependencies to use the basic parsing capability, only that you have a version of PHP 5.x with SimpleXML and either fopen, file_get_contents or CURL enabled. The demonstration of using the "simple_html_dom" to also parse HTML recommendations is merely for completeness' sake and used as a fallback if XML or JSON parsing.